requirements.txt
alembic==1.13.1
amqp==5.2.0
apispec==6.4.0
async-timeout==4.0.3
backports.zoneinfo==0.2.1
billiard==4.2.0
blinker==1.7.0
celery==5.3.5
cffi==1.16.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==41.0.7
Flask==3.0.0
flask-marshmallow==1.1.0
Flask-Migrate==4.0.5
flask-smorest==0.42.3
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
humanize==4.9.0
importlib-metadata==7.0.1
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.3
kombu==5.3.5
Mako==1.3.0
MarkupSafe==2.1.3
marshmallow==3.20.2
marshmallow-sqlalchemy==0.30.0
packaging==23.2
prometheus-client==0.19.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycparser==2.21
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
redis==5.0.1
six==1.16.0
SQLAlchemy==2.0.25
tornado==6.4
typing_extensions==4.9.0
tzdata==2023.4
vine==5.1.0
watchdog==3.0.0
wcwidth==0.2.13
webargs==8.4.0
Werkzeug==3.0.1
zipp==3.17.0
